[pom] Require maven 3.3.9
This most likely actually requires even newer, but still stating 3.2.5 which is extremely old is not ideal.  No one really should use less than 3.8.1 due to vulnerabilities anyways.  So this is not exactly what we want yet but is moving it forwards to block attempted old usage we don't even test for.