[pom] Fix groovy for invoker plugin (still on 3 and breaking changes..)
was not actually overriding groovy to latest.  this was missed when moving from groovy 3 to 4.  We still need 3 with invoker plugin.  The current release of invoker with a breaking change on groovy exposed this and didn't have much in the logs to help.  Basically they dropped all but groovy, json, and xml.  We need nio as well.