Validate Spotbugs report even if only test sources exist
If you use this plugin in a project that only has test sources (source
files in the folder `src/test/`) and no source files in the main folder
(`src/main/`), then it'd not validate the report generated by spotbugs.
Instead it would only log 0 bugs.

Instead of verifying only the folder for class files (main folder), we
also check if there are test class files (only if the option
`includeTests` is enabled). If there are test files, validation of the
report takes place.

Fixes #454