[rework] Before applying this will rework original as it should allow multiple reports
since spotbugs 4.5.0, we can run multiple reports so we should not limit to sarif or xml but rather open that up.  The remainder of this tries to do similar that was done with those but just keeps the temp name for spotbugs html report which does work after fixing this up.  Previously it was full of xml content as only thing done was to rename file.