Feature/generate html reports (#161)
* Add option for generating HTML reports

* Try to add integration test

* Use @pom.version instead of @project.version

* Fix eol marker

* [ci] Adjust formatting to that of original

* [IT] Remove commented out code

* [fix] Correct merge resolution mistake

* [rework] Before applying this will rework original as it should allow multiple reports

since spotbugs 4.5.0, we can run multiple reports so we should not limit to sarif or xml but rather open that up.  The remainder of this tries to do similar that was done with those but just keeps the temp name for spotbugs html report which does work after fixing this up.  Previously it was full of xml content as only thing done was to rename file.

* [ci] Add javadoc for html report

---------

Co-authored-by: Hielke de Vries <hielke.de.vries@ing.com>
Co-authored-by: Jeremy Landis <jeremylandis@hotmail.com>