[pom] Declare proper jakarta.inject-api 1.0.5
maven is using wrong one, the legacy oracle one.  It was long changed to jakarta under javax namespace, The original is missing meta data and not suggested to be used.  I did raise a ticket with maven on this some time ago but it got no where, so since called out as not defined, I'll define the right one and exclude the other.  In the end its the same but at least supported variation even though javax namespace is obsolete.