[pom] ASM bom was new recently, previously these needed overridden, will just use bom now
Spotbugs itself went dormant for a long time and now there are only 3 active members that includes myself.  So while controlling asm is less important, it is still very important as groovy overrides asm and so does a number of other items in general and at any given time we may not release spotbugs in timely fashion as the plugins are released far faster.  This will get it off the issue with analyzer though.