Cleanup base violation check mojo
Remove all items that do not belong in this class.  This part looks only at the results, it does not need the information for the reports.  The runs execute 'spotbugs' typically then verify the results.  If the verify one is used, it doesn't execute spotbugs only checks pre-existing.  In either case, most of this was just duplicated from the original mojo and unnecessary.

Also cleanup exit when report directory does not exist to reduce nesting