Fix usage of 'debug' flag vs 'logging debug flag' as they are not the same
In multiple spots debug logging flags were used as if debugging spotbugs process, that was invalid.  Use proper debug flag to make it easier for users to debug.  None of that had direct relationship to logging outside of GUI but in that case, move logging to info level as it is unrelated to logging in that regard based on usage (although it logs).