[gha] Run it on all platforms but slip to save time
maven 3.6.3 -> run on windows
maven 4.0.0 -> run on macos
maven latest -> run on linux

An issue with normalization was detected on windows but not seen on github, this will ensure we detect such issues by exercising integration tests on all platforms.  Given how long integration tests run, this is a trade off to keep same general speed in place.